Implement teleporting vehicles to starting positions, fix bug with throttle input
If player hold throttle on countdown, it would not be read properly, it never happens in the first race, I couldn't really identify the problem - something with changing the input, but find a workaround, now bug still occurs if players holds W through the whole race finished widget. Probably we need to flush input somehow.